requirements.txt
PyQt5==5.15.11
pygetwindow==0.0.9
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.4.0
